requirements.txt
alembic==1.16.1
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.47.0
greenlet==3.2.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.5
onnxruntime==1.22.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
playsound==1.2.2
protobuf==6.31.0
pymongo==4.13.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-telegram-bot==22.1
pyttsx3==2.98
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
telegram==0.0.1
tifffile==2025.5.10
torch==2.7.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
XlsxWriter==3.2.3
gunicorn
